+"Setup Git VCS and change UART4 to USART1"
Changes in this commit include setting up Git as the version control system and changing the USART from UART4 to USART1 in the "usart.c" file and "LiangShanPai.ioc" file. This was done to align with hardware changes and to ensure smooth collaboration through Git. Also, new mappings of directories and relevant files are added to the ".mxproject" file for efficient tracking of changes.